+++ client/src/scss/signup.scss
@@ -1,5 +1,5 @@
 .base-container {
-  width: 100%;
+  width: 50%;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -40,10 +40,6 @@
           border-radius: 4px;
           margin-bottom: 31px;
           transition: all 250ms ease-in-out;
-          &:hover {
-            //background-color: #ffffff;
-            //box-shadow: 0px 0px 14px 0.3px #0e81ce96;
-          }
 
           &:focus {
             outline: none;
